Make use of dependency injection in seeder
